web: consistent buttons between two pages.
Tana pointed out that the language and functionality for password
recovery between these two pages is very inconsistent, so I made
them so. I revised the buttons on the ExpandedUserList and UserViewPage
so that they have the same labels for the same functionality, added
the Send User A Password Reset in Email button to the UserViewPage,
and extracted the "Put something in the clipboard and tell the
customer about it" functionality from the AkTokenButton and used
it for the View Password Reset Link feature, since that's what we
expect a customer to do with that URL anyway.