web: ferreting out all the bugs
This commit includes a new, wholly shadow-dommed component for `<select multiple>`, with a slightly new and improved
encapsulation of the HorizontalInputControl.  This is *not* how I wanted to do this; it's starting to sprawl with
a lot of different "experiments" in input types, but this, at least, works.

It fixes the "multiple display" issue with the SCIM input, too.  It's a start.